requirements.txt
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
dnspython==2.2.1
filelock==3.6.0
Flask==2.1.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
password-strength==0.0.3.post2
phonenumbers==8.12.46
py3-validate-email==1.0.5
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
